+++ libraries/OlcbCommonCAN/OpenLcbCanBuffer.h
@@ -75,7 +75,7 @@ class EventID;
   void getEventID(EventID* evt);
   void getNodeID(NodeID* nid);
   
-  bool isVerifyNID(int nida);
+  bool isVerifyNID(uint16_t nida);
   bool isVerifyNIDglobal();
   void setVerifiedNID(NodeID* nid);
   bool isVerifiedNID();
@@ -96,7 +96,7 @@ class EventID;
   // Mask uses an EventID data structure; 1 bit means mask out when routing
   void setProducerIdentifyRange(EventID* eid, EventID* mask);
 
-  bool isIdentifyEvents(int nida);
+  bool isIdentifyEvents(uint16_t nida);
   bool isIdentifyEventsGlobal();
 
   bool isDatagram();
